+++ test/Main.hs
@@ -11,15 +11,19 @@ module Main (main) where
 
 import Test.Tasty (TestTree, defaultMain, testGroup)
 
+import Effectful.Tracing.AttributeSpec qualified as AttributeSpec
 import Effectful.Tracing.CompileTest qualified as CompileTest
 import Effectful.Tracing.ConcurrentSpec qualified as ConcurrentSpec
 import Effectful.Tracing.FuzzSpec qualified as FuzzSpec
+import Effectful.Tracing.IdsSpec qualified as IdsSpec
 import Effectful.Tracing.InMemorySpec qualified as InMemorySpec
+import Effectful.Tracing.LifecycleSpec qualified as LifecycleSpec
 import Effectful.Tracing.NoOpSpec qualified as NoOpSpec
 import Effectful.Tracing.PrettyPrintSpec qualified as PrettyPrintSpec
 import Effectful.Tracing.PropagationSpec qualified as PropagationSpec
 import Effectful.Tracing.PropertySpec qualified as PropertySpec
 import Effectful.Tracing.SamplerSpec qualified as SamplerSpec
+import Effectful.Tracing.TypesSpec qualified as TypesSpec
 
 #ifdef OTEL
 import Effectful.Tracing.OpenTelemetrySpec qualified as OpenTelemetrySpec
@@ -39,9 +43,13 @@ main =
     ( testGroup
         "effectful-tracing"
         ( [ PropertySpec.tests
+          , TypesSpec.tests
+          , AttributeSpec.tests
+          , IdsSpec.tests
           , CompileTest.tests
           , NoOpSpec.tests
           , InMemorySpec.tests
+          , LifecycleSpec.tests
           , PrettyPrintSpec.tests
           , SamplerSpec.tests
           , ConcurrentSpec.tests